[FO-2389] flytter kanSendeMelding ut i egen reackt hook
