+++ src/view/dialog/NyDialog.tsx
@@ -15,6 +15,7 @@ import { VenstreChevron } from 'nav-frontend-chevron';
 import { Link } from 'react-router-dom';
 
 import './NyDialog.less';
+import useKansendeMelding from '../../utils/UseKanSendeMelding';
 
 const AlertStripeFeilVisible = visibleIfHoc(AlertStripeFeil);
 
@@ -43,9 +44,20 @@ function NyDialog() {
     const dialoger = useDialogContext();
     const bruker = useUserInfoContext();
     const history = useHistory();
+    const kanSendeMeldign = useKansendeMelding();
 
     const [ferdigBehandlet, setFerdigBehandlet] = useState(true);
     const [venterPaSvar, setVenterPaSvar] = useState(false);
+
+    if (!kanSendeMeldign) {
+        return (
+            <div className="dialog dialog-new">
+                Du kan ikke sende dialog
+                <br />
+                Se feilmelding over
+            </div>
+        );
+    }
 
     function handleSubmit(event: FormEvent) {
         event.preventDefault();
@@ -104,43 +116,40 @@ function NyDialog() {
     };
 
     return (
-        <>
-            <div className="dialog dialog-new">
-                <div className="dialog-new__header">
-                    <Link to="/" className="tilbake-til-oversikt">
-                        <VenstreChevron stor className="tilbake-til-oversikt__pilknapp" />
-                        Oversikt
-                    </Link>
-                </div>
-                <form onSubmit={handleSubmit} noValidate className="dialog-new__form">
-                    <Innholdstittel className="dialog-new__tittel">Ny dialog</Innholdstittel>
-                    <Normaltekst className="dialog-new__infotekst">
-                        Her kan du skrive til din veileder om arbeid og oppfølging. Du vil få svar i løpet av noen
-                        dager.
-                    </Normaltekst>
-                    <Valideringsboks tema={tema} melding={melding} triedToSubmit={triedToSubmit} />
-                    <Input
-                        id="temaIn"
-                        className="dialog-new__temafelt"
-                        label={'Tema:'}
-                        placeholder="Skriv her"
-                        {...tema.input}
-                        disabled={submitting}
-                    />
-                    <HenvendelseInput melding={melding} submitting={submitting} />
-                    <DialogCheckboxesVisible
-                        toggleFerdigBehandlet={toggleFerdigBehandlet}
-                        toggleVenterPaSvar={toggleVenterPaSvar}
-                        ferdigBehandlet={ferdigBehandlet}
-                        venterPaSvar={venterPaSvar}
-                        visible={bruker!.erVeileder}
-                    />
-                    <AlertStripeFeilVisible visible={submitfeil}>
-                        Det skjedde en alvorlig feil. Prøv igjen senere
-                    </AlertStripeFeilVisible>
-                </form>
+        <div className="dialog dialog-new">
+            <div className="dialog-new__header">
+                <Link to="/" className="tilbake-til-oversikt">
+                    <VenstreChevron stor className="tilbake-til-oversikt__pilknapp" />
+                    Oversikt
+                </Link>
             </div>
-        </>
+            <form onSubmit={handleSubmit} noValidate className="dialog-new__form">
+                <Innholdstittel className="dialog-new__tittel">Ny dialog</Innholdstittel>
+                <Normaltekst className="dialog-new__infotekst">
+                    Her kan du skrive til din veileder om arbeid og oppfølging. Du vil få svar i løpet av noen dager.
+                </Normaltekst>
+                <Valideringsboks tema={tema} melding={melding} triedToSubmit={triedToSubmit} />
+                <Input
+                    id="temaIn"
+                    className="dialog-new__temafelt"
+                    label={'Tema:'}
+                    placeholder="Skriv her"
+                    {...tema.input}
+                    disabled={submitting}
+                />
+                <HenvendelseInput melding={melding} submitting={submitting} />
+                <DialogCheckboxesVisible
+                    toggleFerdigBehandlet={toggleFerdigBehandlet}
+                    toggleVenterPaSvar={toggleVenterPaSvar}
+                    ferdigBehandlet={ferdigBehandlet}
+                    venterPaSvar={venterPaSvar}
+                    visible={bruker!.erVeileder}
+                />
+                <AlertStripeFeilVisible visible={submitfeil}>
+                    Det skjedde en alvorlig feil. Prøv igjen senere
+                </AlertStripeFeilVisible>
+            </form>
+        </div>
     );
 }
